Issue #11: makefile, coverage test and all functions that need to be test it
