+++ makefile
@@ -1,0 +1,26 @@
+#! usr/bin/env bash
+
+TEST_DIR=tests/
+
+.PHONY: preInstall
+preInstall:
+	pip install --upgrade pip
+	pip install -r requirements.txt
+
+.PHONY: test
+test: cleanReport
+	pytest $(TEST_DIR)
+
+
+.PHONY: cleanReport
+cleanReport:
+	rm -rf .coverage
+
+
+.PHONY: openReport
+openReport: cleanReport
+	coverage run --branch -m unittest discover --pattern '*_test.py' --start-directory $(TEST_DIR)
+	coverage report --show-missing --include vandor_rename.py
+	coverage html --title 'Tests Coverage Report' --include vandor_rename.py
+	open htmlcov/index.html
+
