+++ requirements.txt
@@ -1,0 +1,2 @@
+pytest==5.2.0
+coverage==4.5.4
