--- tests/requirements.txt
@@ -1,1 +1,0 @@
-pytest==5.2.0
